2.7.1: 회귀 테스트 게이트 확충 + ARM64 이식성
테스트 (4 → 64개):
- tests/test_csv_serialize.py: log_csv/format_csv 계약 테스트 — f-string 바이트
  동일성(타입 매트릭스 + 시드 고정 300회 랜덤), 스키마/에러 경로(예외 타입 매핑),
  4096B 라인 버퍼 경계 7종(정확히 채우기 성공, 콤마/nan/to_chars 오버플로우 —
  v2.5.0 bounds check 회귀 가드), 파일 출력 라운드트립, csv_time_formatter 패턴
- tests/test_logger_behavior.py: 파일 출력 내용/레벨 필터링/flush_on 즉시 가시성,
  str-only 계약(bytes TypeError), 레지스트리 get/drop/재사용, async 모드
  (서브프로세스 격리, 500라인 드레인 검증), 유니코드, 싱크류 스모크

ARM64 이식성 (별도 코드리뷰로 검증, LOW 8건 외 이슈 없음):
- -march=native/-mtune=native를 x86 전용으로 게이트, aarch64/arm64는 -mcpu=native
  (일부 ARM 툴체인은 -march=native 거부 — 빌드 차단급)
- -DSPDLOG_FWRITE_UNLOCKED를 Darwin 제외 게이트 (fwrite_unlocked는 GNU 확장)
- dist_sink의 비표준 uint → size_t
- NO_ATOMIC_LEVELS·to_chars 바이트 동일성·char 부호 등은 ARM64 약한 메모리
  모델에서도 안전함을 리뷰로 확인 (셋업 시점 레벨 설정 패턴 전제)

검증: 전체 64 테스트 통과, format_csv 바이트 동일성 15케이스, x86 재빌드 확인

+++ src/pyspdlog.cpp
@@ -381,7 +381,7 @@ public:
     dist_sink(std::vector<Sink> sinks)
     {
         std::vector<spd::sink_ptr> sink_vec;
-        for (uint i =0; i<sinks.size(); i++)
+        for (size_t i = 0; i < sinks.size(); i++)
         {
             sink_vec.push_back(sinks.at(i).get_sink());
         }
@@ -400,7 +400,7 @@ public:
     void set_sinks(std::vector<Sink> sinks)
     {
         std::vector<spd::sink_ptr> sink_vec;
-        for (uint i =0; i<sinks.size(); i++)
+        for (size_t i = 0; i < sinks.size(); i++)
         {
             sink_vec.push_back(sinks.at(i).get_sink());
         }
